requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pynostr
pytest
alembic